backend/requirements.txt
Django==5.2.5
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-decouple==3.8 